Debugging Cadence

+++ include/smartbike_parameters.h
@@ -35,7 +35,7 @@ class userParameters
     float       getIncline()                 {return incline;}
     int         getSimulatedWatts()          {return simulatedWatts;}
     int         getSimulatedHr()             {return simulatedHr;}
-    int         getSimulatedCad()            {return simulatedCad;}
+    float         getSimulatedCad()            {return simulatedCad;}
     float       getInclineStep()             {return inclineStep;}
     int         getShiftStep()               {return shiftStep;}
     float       getInclineMultiplier()       {return inclineMultiplier;}
@@ -54,7 +54,7 @@ class userParameters
     void    setIncline(float inc)            {incline = inc;}
     void    setSimulatedWatts(int w)         {simulatedWatts = w;}
     void    setSimulatedHr(int hr)           {simulatedHr = hr;}
-    void    setSimulatedCad(int cad)         {simulatedCad = cad;}
+    void    setSimulatedCad(float cad)         {simulatedCad = cad;}
     void    setInclineStep(float is)         {inclineStep = is;}
     void    setShiftStep(int ss)             {shiftStep = ss;}
     void    setInclineMultiplier(float im)   {inclineMultiplier = im;}
